Add DataTriaseIgd model
